Add simple AST traversal
This code, inspired by the defunct NPComp, walks the python AST to
emit MLIR ops.  This is intended for kernel code in Iron, and is
complementary to the more MLIR-centric generation elsewhere.
This is a proof of concept only, several bits are needed to making
it work in practice:
1) A more complete type system, probably based off of NPComps BasicPy
dialect and some type inference
2) Need to walk python for loops to determine liveins/liveouts/recurrences
3) Integrate with python type annotations at the boundaries
4) bindings from tensor ops into vector types
5) Integration with the rest of Iron  (A different decorator?)